% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{BSnorm}
\alias{BSnorm}
\title{Normalized Brier Score}
\usage{
BSnorm(probabilities, y)
}
\arguments{
\item{probabilities}{matrix of predicted probabilities for each class (classes in columns, observations in rows)}

\item{y}{matrix of true values for y; each column contains a class ordered in the same order as probabilities}
}
\value{
Normalized Brier Score
}
\description{
Normalized Brier Score
}
