% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_convergence.R
\name{calculate_convergence}
\alias{calculate_convergence}
\title{Calculates convergence based on NMSE}
\usage{
calculate_convergence(OOB_err, OOB_weights, ximp_new, ximp_old, xmis)
}
\arguments{
\item{OOB_err}{dataframe containing OOB errors for each iteration.}

\item{OOB_weights}{vector of weights in the same format as for the \code{missForest} function.}

\item{ximp_new}{imputed dataframe of current iteration}

\item{ximp_old}{imputed dataframe of previous iteration}

\item{xmis}{the original dataframe containing missing values.}
}
\value{
A list with elements
    \item{\code{converged}}{boolean indicating if the algorithm has converged (TRUE) or not (FALSE)}
    \item{\code{measure_old}}{old measure: TODO: will this work?}
    \item{\code{measure_new}}{new measure: TODO: will this work?}
}
\description{
Calculates convergence based on NMSE. TODO: blabla
}
