% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predictor_matrix.R
\name{check_predictor_matrix}
\alias{check_predictor_matrix}
\title{Performs checks on a custom predictor matrix}
\usage{
check_predictor_matrix(predictor_matrix, data, verbose = TRUE)
}
\arguments{
\item{predictor_matrix}{custom predictor matrix}

\item{data}{dataframe to be imputed}

\item{verbose}{if TRUE, returns human readable message for the checks performed.}
}
\value{
No return value, only verifies the validity the provided predictor matrix
}
\description{
Performs a number of checks on a custom predictor matrix and returns error if the matrix is
not in a valid format.
}
