% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_imputation_error.R
\name{evaluate_imputation_error}
\alias{evaluate_imputation_error}
\title{Evaluate the imputation error when true values are known.}
\usage{
evaluate_imputation_error(ximp, xmis, xtrue)
}
\arguments{
\item{ximp}{imputed dataframe}

\item{xmis}{original dataframe with missing values}

\item{xtrue}{true dataframe (or any matrix to be compared with ximp)}
}
\value{
named vector of errors
}
\description{
Evaluate the imputation error when true values are known. TODO
}
