% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary.R
\name{make_binary}
\alias{make_binary}
\title{Gets a factor variable, returns a matrix of binary variables}
\usage{
make_binary(x)
}
\arguments{
\item{x}{a vector of type factor}
}
\value{
a matrix of binary variables
}
\description{
Gets a factor variable, returns a matrix of binary variables
}
