% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mer}
\alias{mer}
\title{Missclassification error rate - ER}
\usage{
mer(prediction, y)
}
\arguments{
\item{y}{matrix of true values for y; each column contains a class ordered in the same order as probabilities}

\item{probabilities}{matrix of class predictions for each class (classes in columns, observations in rows)}
}
\value{
Missclassification error rate
}
\description{
Missclassification error rate - ER
}
