% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missForest_old.R
\name{missForest_old}
\alias{missForest_old}
\title{Title}
\usage{
missForest_old(
  xmis,
  maxiter = 10,
  ntree = 500,
  variablewise = FALSE,
  decreasing = FALSE,
  verbose = FALSE,
  mtry = floor(sqrt(ncol(xmis))),
  replace = TRUE,
  class.weights = NULL,
  cutoff = NULL,
  strata = NULL,
  sample.fraction = NULL,
  maxnodes = NULL,
  xtrue = NA
)
}
\arguments{
\item{xmis}{foo}

\item{maxiter}{foo}

\item{ntree}{foo}

\item{variablewise}{foo}

\item{decreasing}{foo}

\item{verbose}{foo}

\item{mtry}{foo}

\item{replace}{foo}

\item{class.weights}{foo}

\item{cutoff}{foo}

\item{strata}{foo}

\item{sample.fraction}{foo}

\item{maxnodes}{foo}

\item{xtrue}{foo}
}
\value{
foo
}
\description{
Title
}
