% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missForestPredict_old.R
\name{missForestPredict_old}
\alias{missForestPredict_old}
\title{Imputes a new dataframe based on the missForest models}
\usage{
missForestPredict_old(missForestObj, newdata)
}
\arguments{
\item{missForestObj}{missForest object as returned by the missForest function}

\item{newdata}{new data to impute. The column names should be the same as in the imputation model}
}
\value{
an imputed dataframe
}
\description{
Imputes a new dataframe based on the missForest models
}
