% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixError.R
\name{mixError}
\alias{mixError}
\title{Title}
\usage{
mixError(ximp, xmis, xtrue, variableWise = FALSE)
}
\arguments{
\item{ximp}{imputed matrix / dataframe}

\item{xmis}{original matrix / dataframe with missing values}

\item{xtrue}{true matrix / dataframe (or any matrix to be compared with ximp)}

\item{variableWise}{logical; should the errors be returned for each variable (TRUE) or aggregated by variable type (numeric / factor)}
}
\value{
named vector of errors
}
\description{
Title
}
