% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mse}
\alias{mse}
\title{Mean Square Error - MSE}
\usage{
mse(preds, y)
}
\arguments{
\item{preds}{vector of predictions}

\item{y}{vector of true values for y}
}
\value{
Mean Square Error - MSE
}
\description{
Mean Square Error - MSE
}
