% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{nmse}
\alias{nmse}
\title{Normalized Mean Square Error - NMSE}
\usage{
nmse(preds, y)
}
\arguments{
\item{preds}{vector of predictions}

\item{y}{vector of true values for y}
}
\value{
Normalized Mean Square Error - NMSE
}
\description{
Normalized Mean Square Error - NMSE
}
