% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodNA.R
\name{prodNA}
\alias{prodNA}
\title{Title}
\usage{
prodNA(x, noNA = 0.1)
}
\arguments{
\item{x}{matrix / dataframe}

\item{noNA}{percentage of NA on each column}
}
\value{
matrix / dataframe with missing values
}
\description{
Title
}
