% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_NA.R
\name{produce_NA}
\alias{produce_NA}
\title{Produces a dataframe with missing values}
\usage{
produce_NA(x, proportion = 0.3)
}
\arguments{
\item{x}{dataframe}

\item{proportion}{percentage of NA on each column}
}
\value{
dataframe with missing values
}
\description{
Produces a dataframe with missing values
}
