% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_imputation_error.R
\name{evaluate_imputation_error}
\alias{evaluate_imputation_error}
\title{Title}
\usage{
evaluate_imputation_error(ximp, xmis, xtrue)
}
\arguments{
\item{ximp}{imputed matrix / dataframe}

\item{xmis}{original matrix / dataframe with missing values}

\item{xtrue}{true matrix / dataframe (or any matrix to be compared with ximp)}
}
\value{
named vector of errors
}
\description{
Title
}
