% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_NA.R
\name{produce_NA}
\alias{produce_NA}
\title{Produces a dataframe with missing values}
\usage{
produce_NA(x, proportion = 0.1)
}
\arguments{
\item{x}{a dataframe}

\item{proportion}{proportion of missing values to be produced; a vector of size \code{ncol(x)}
or a single value to be applied to all variables in the dataframe.}
}
\value{
dataframe with missing values.
}
\description{
Produces a dataframe with missing values.
For each variable, a random sample of observations of size \code{proportion * nrow(x)} will be set to NA.
}
